requirements.txt
python-telegram-bot
pytesseract
Pillow
pdf2image